user-token part done with connection of user to expense database

+++ controllers/expenseController.js
@@ -12,39 +12,29 @@ exports.getExpense=async (req,res,next)=>{
     }
 }
 exports.getAllExpense=async (req,res,next)=>{
-    // DailyExpense.getAllExpense().then((response)=>{
-    //     res.send(response);
-    // })
     try{
-        const expense = await DailyExpense.findAll()
+        console.log(req.user.id)
+        const expense = await DailyExpense.findAll({where:{userId:req.user.id}})
         res.json(expense);
 
     }catch(error){
         console.log(error)
+        return res.status(500).json({error:error,success:false})
     }
 }
 
 exports.addExpense=async (req,res,next)=>{
-    // const{amount,description,category}=req.body;
-
-    // const dailyExpense=new DailyExpense(null,amount,description,category);
-
-    // dailyExpense.addExpense().then((response)=>{
-    //     console.log(response);
-    //     res.send(response);
-    // }).catch((err)=>{
-    //     res.status(500).send("Error occurred while adding the expense.");
-    // })
     try{
         const{amount,description,category}=req.body;
         await DailyExpense.create({
             amount:amount,
             description:description,
-            category:category
+            category:category,
+            userId:req.user.id
         }).then((response)=>{
             res.status(200).json({message:"Expenses added"});
         }).catch((err)=>{
-            console.log(err);
+            res.status(500).then("Error occured while adding the expenses")
         })
     }catch(err){
         console.log(err)
@@ -54,22 +44,13 @@ exports.addExpense=async (req,res,next)=>{
 
 }
 exports.deleteExpense=async (req,res,next)=>{
-    // conso
-    // const dailyExpense = new DailyExpense(req.body.id,null,null,null);
-    // dailyExpense.deleteExpense().then((response)=>{
-    //     console.log("delte done")
-    //     res.status(200).send(response)
-    // }).catch((err)=>{
-    //     console.log("nodDete")
-    //     res.status(500).send("Error occurred while deleting the expense.");
-
-
-    // })
+    
     const id = req.params.id;
-    console.log(id)
+    console.log(id);
+    console.log(req.user.id);
     try{
         const expense = await DailyExpense.findByPk(id);
-        await DailyExpense.destroy({where:{id:id}}).then((response)=>{
+        await DailyExpense.destroy({where:{id:id,userId:req.user.id}}).then((response)=>{
             res.status(200).json({message:"Deleted successfully"})
         }).catch((err)=>{
             console.log(err);
